+++ src/main/resources/db/migrations/V14__create_stats_table.sql
@@ -1,6 +1,6 @@
 create table stats (
         num_items INT,
         time INT,
-        latitude FLOAT NOT NULL,
-        longitude FLOAT NOT NULL
+        latitude DOUBLE NOT NULL,
+        longitude DOUBLE NOT NULL
 );
