+++ src/main/resources/db/migrations/V6__create_crowd_product_prices_table.sql
@@ -1,7 +1,7 @@
 create table crowd_product_prices (
     barcode VARCHAR(255),
     price DOUBLE NOT NULL,
-    latitude FLOAT NOT NULL,
-    longitude FLOAT NOT NULL,
+    latitude DOUBLE NOT NULL,
+    longitude DOUBLE NOT NULL,
     PRIMARY KEY (barcode, latitude, longitude)
 );
